Convert assignments page and CycleSelector to server components + actions
Major API simplification breakthrough:
- Converted /evaluations/assignments from client component + API calls to server component + server actions
- Fixed CycleSelector component to use server-provided data instead of /api/admin/cycles endpoints
- Added "Completed" badge for finished evaluations instead of confusing overdue warnings

Changes:
• Server Components & Actions:
  - /evaluations/assignments: Direct DB queries + server actions for bulk assignment operations
  - /dashboard: Extended to fetch and provide cycles data to CycleSelector component
  - Created server actions for cycle management (close/reopen/create)

• UI Improvements:
  - Added green "Completed" badge in evaluation page for completed evaluations
  - Fixed company-wide items visibility in assignments page
  - Improved error handling with proper transitions and loading states

• API Elimination:
  - Removed dependency on /api/manager/team-assignments (GET) - now server component
  - Fixed CycleSelector API dependency issue that was causing "Failed to load cycle" errors
  - Maintained all existing functionality while eliminating API complexity

Progress: 11 of 21 API endpoints eliminated (52% reduction achieved\!)
