+++ src/app/evaluations/assignments/actions.ts
@@ -1,0 +1,244 @@
+'use server'
+
+import { auth } from '@/auth'
+import { prisma } from '@/lib/prisma-client'
+import { revalidatePath } from 'next/cache'
+
+// Server action to assign evaluation items to employees
+export async function assignItemsToEmployees(itemId: string, employeeIds: string[]) {
+  try {
+    const session = await auth()
+    
+    if (!session?.user?.id) {
+      return { success: false, error: 'Unauthorized' }
+    }
+
+    const userRole = session.user.role
+    const companyId = session.user.companyId
+
+    // Only managers and HR can assign items
+    if (userRole !== 'manager' && userRole !== 'hr') {
+      return { success: false, error: 'Access denied - Manager or HR role required' }
+    }
+
+    // Create assignments for each employee
+    const assignments = employeeIds.map(employeeId => ({
+      evaluationItemId: itemId,
+      employeeId,
+      companyId,
+      assignedBy: session.user.id,
+      assignedAt: new Date()
+    }))
+
+    await prisma.evaluationItemAssignment.createMany({
+      data: assignments,
+      skipDuplicates: true // Prevent duplicate assignments
+    })
+
+    revalidatePath('/evaluations/assignments')
+    return { success: true }
+
+  } catch (error) {
+    console.error('Error assigning items:', error)
+    return { success: false, error: 'Failed to assign items' }
+  }
+}
+
+// Server action to unassign evaluation items from employees
+export async function unassignItemsFromEmployees(itemId: string, employeeIds: string[]) {
+  try {
+    const session = await auth()
+    
+    if (!session?.user?.id) {
+      return { success: false, error: 'Unauthorized' }
+    }
+
+    const userRole = session.user.role
+    const companyId = session.user.companyId
+
+    // Only managers and HR can unassign items
+    if (userRole !== 'manager' && userRole !== 'hr') {
+      return { success: false, error: 'Access denied - Manager or HR role required' }
+    }
+
+    // Delete assignments
+    await prisma.evaluationItemAssignment.deleteMany({
+      where: {
+        evaluationItemId: itemId,
+        employeeId: { in: employeeIds },
+        companyId
+      }
+    })
+
+    revalidatePath('/evaluations/assignments')
+    return { success: true }
+
+  } catch (error) {
+    console.error('Error unassigning items:', error)
+    return { success: false, error: 'Failed to unassign items' }
+  }
+}
+
+// Server action to create new evaluation item
+export async function createEvaluationItem(formData: {
+  title: string
+  description: string
+  type: 'okr' | 'competency'
+  level: 'company' | 'department' | 'manager'
+  evaluationDeadline?: string
+}) {
+  try {
+    const session = await auth()
+    
+    if (!session?.user?.id) {
+      return { success: false, error: 'Unauthorized' }
+    }
+
+    const userId = session.user.id
+    const userRole = session.user.role
+    const companyId = session.user.companyId
+
+    // Only managers and HR can create items
+    if (userRole !== 'manager' && userRole !== 'hr') {
+      return { success: false, error: 'Access denied - Manager or HR role required' }
+    }
+
+    const { title, description, type, level, evaluationDeadline } = formData
+
+    if (!title?.trim() || !description?.trim()) {
+      return { success: false, error: 'Title and description are required' }
+    }
+
+    // Permission check
+    if (userRole === 'manager' && level === 'company') {
+      return { success: false, error: 'Only HR can create company-level items' }
+    }
+
+    // Validate deadline if provided
+    let deadlineDate = null
+    if (evaluationDeadline) {
+      deadlineDate = new Date(evaluationDeadline)
+      if (isNaN(deadlineDate.getTime())) {
+        return { success: false, error: 'Invalid deadline date format' }
+      }
+      
+      const now = new Date()
+      const oneHourFromNow = new Date(now.getTime() + 60 * 60 * 1000)
+      if (deadlineDate <= oneHourFromNow) {
+        return { success: false, error: 'Deadline must be at least 1 hour in the future' }
+      }
+    }
+
+    // Get user info for assignedTo field
+    const user = await prisma.user.findUnique({
+      where: { id: userId },
+      select: { department: true }
+    })
+
+    let assignedTo = null
+    if (level === 'department' && user?.department) {
+      assignedTo = user.department
+    } else if (level === 'manager') {
+      assignedTo = userId
+    }
+
+    // Get the next sort order
+    const lastItem = await prisma.evaluationItem.findFirst({
+      where: { companyId },
+      orderBy: { sortOrder: 'desc' }
+    })
+    const sortOrder = (lastItem?.sortOrder || 0) + 1
+
+    // Create the item
+    await prisma.evaluationItem.create({
+      data: {
+        companyId,
+        title: title.trim(),
+        description: description.trim(),
+        type,
+        level,
+        createdBy: userId,
+        assignedTo,
+        sortOrder,
+        active: true,
+        evaluationDeadline: deadlineDate,
+        deadlineSetBy: deadlineDate ? userId : null
+      }
+    })
+
+    revalidatePath('/evaluations/assignments')
+    revalidatePath('/dashboard/company-items')
+    return { success: true }
+
+  } catch (error) {
+    console.error('Error creating evaluation item:', error)
+    return { success: false, error: 'Failed to create evaluation item' }
+  }
+}
+
+// Server action to update evaluation item
+export async function updateEvaluationItem(itemId: string, formData: {
+  title: string
+  description: string
+  evaluationDeadline?: string
+}) {
+  try {
+    const session = await auth()
+    
+    if (!session?.user?.id) {
+      return { success: false, error: 'Unauthorized' }
+    }
+
+    const userId = session.user.id
+    const userRole = session.user.role
+    const companyId = session.user.companyId
+
+    // Only managers and HR can update items
+    if (userRole !== 'manager' && userRole !== 'hr') {
+      return { success: false, error: 'Access denied - Manager or HR role required' }
+    }
+
+    const { title, description, evaluationDeadline } = formData
+
+    if (!title?.trim() || !description?.trim()) {
+      return { success: false, error: 'Title and description are required' }
+    }
+
+    // Validate deadline if provided
+    let deadlineDate = null
+    if (evaluationDeadline) {
+      deadlineDate = new Date(evaluationDeadline)
+      if (isNaN(deadlineDate.getTime())) {
+        return { success: false, error: 'Invalid deadline date format' }
+      }
+      
+      const now = new Date()
+      const oneHourFromNow = new Date(now.getTime() + 60 * 60 * 1000)
+      if (deadlineDate <= oneHourFromNow) {
+        return { success: false, error: 'Deadline must be at least 1 hour in the future' }
+      }
+    }
+
+    // Update the item
+    await prisma.evaluationItem.update({
+      where: { 
+        id: itemId,
+        companyId // Ensure user can only update items from their company
+      },
+      data: {
+        title: title.trim(),
+        description: description.trim(),
+        evaluationDeadline: deadlineDate,
+        deadlineSetBy: deadlineDate ? userId : null
+      }
+    })
+
+    revalidatePath('/evaluations/assignments')
+    revalidatePath('/dashboard/company-items')
+    return { success: true }
+
+  } catch (error) {
+    console.error('Error updating evaluation item:', error)
+    return { success: false, error: 'Failed to update evaluation item' }
+  }
+}
